+++ @remirror/extension-code-block/src/code-block-utils.ts
@@ -12,8 +12,8 @@ import {
   NodeType,
   NodeTypeParams,
   NodeWithPosition,
-  PMNodeParams,
   PosParams,
+  ProsemirrorNodeParams,
   TextParams,
   uniqueArray,
 } from '@remirror/core';
@@ -134,7 +134,7 @@ export const posWithinRange = ({ from, to, pos }: PosWithinRangeParams) => from 
 export const lengthHasChanged = <GType>(prev: ArrayLike<GType>, next: ArrayLike<GType>) =>
   next.length !== prev.length;
 
-export interface NodeInformation extends NodeTypeParams, FromToParams, PMNodeParams, PosParams {}
+export interface NodeInformation extends NodeTypeParams, FromToParams, ProsemirrorNodeParams, PosParams {}
 
 /**
  * Retrieves helpful node information from the current state.
@@ -184,17 +184,6 @@ export const updateNodeAttrs = (type: NodeType) => (attrs?: Attrs): CommandFunct
 
   return true;
 };
-
-// interface FormatCodeBlockTransactionParams extends FormatCodeBlockFactoryParams, RangeParams, TransactionParams {}
-// /**
-//  * Updates the provided Transaction with a transformed codeBlock
-//  *
-//  * The transaction is mutated if there is a change and the function returns true to indicate
-//  * that it should be dispatched and false to suggest otherwise.
-//  */
-// const formatCodeBlockTransaction = ({}: FormatCodeBlockTransactionParams): boolean => {
-
-// }
 
 interface FormatCodeBlockFactoryParams
   extends NodeTypeParams,